Add files via upload
I chose the data set called “Crest Sales” as my data to analyze and forecast; which is data set number 563 in the tsdl package found in RStudio packages. This data set contains weekly market share data of Crest sales from January 1958 to April 1963. This data set is interesting to me because I myself am a customer of Crest products such as their toothpaste. So it is interesting to learn more about their market share.